cip-6290 fixed data types that should have been floats

+++ src/ska_mid_cbf_fhs_vcc/vcc_all_bands/vcc_all_bands_dataclasses.py
@@ -10,7 +10,7 @@ from ska_mid_cbf_fhs_common import FhsControllerBaseConfig
 class VCCAllBandsConfigureScanPowerMeterConfig(DataClassJsonMixin):
     """Dataclass representing a power meter configuration as part of the VCC All Bands ConfigureScan input parameter."""
 
-    averaging_time: int
+    averaging_time: float
     flagging: int
 
 
@@ -20,7 +20,7 @@ class VCCAllBandsConfigureScanFSLaneConfig(DataClassJsonMixin):
 
     vlan_id: int
     fs_id: int
-    averaging_time: int
+    averaging_time: float
     flagging: int
 
 
@@ -58,8 +58,8 @@ class VCCAllBandsConfigureVCCBiteNoiseInfoPolarityConfig(DataClassJsonMixin):
     """Dataclass representing Polarity config for Noise Info field in VCC Bite Config"""
 
     seed: int
-    noise_std: int
-    noise_mean: int
+    noise_std: float
+    noise_mean: float
 
 
 @dataclass
